`fbSendRequest` now actually returns `False` if the request could not be send because the connection is closed.
There is no practical way to check if a socket is connected, other than to try to use it and catch an exception if this fails.

HTTP connections can be reused, so they can stay open after a request was sent and a response received. The connect can be closed by the server at any point, which we will only learn about when we try to send another request. This should be handled gracefully by using a different connection, or creating a new one, to send teh request to the server. So, we want to know if we cannot send a request because the connection was closed. In this specific case, the method should return False. It now actually does this.