Same as previous
After a request/response pair has been excahnged, the connection may stay open but can be closed by the server at any point. We will not find out until we try to us it. If we are reusing a connection and cannot send the request because it is closed, we will try reusing another one, or create a new one and send again.